requirements.txt
colorama==0.4.5
pyfiglet==0.7
requests==2.28.1
python-dotenv==0.20.0